Uncomment header include
This header was transitively included by the previous include. We should be explicit with all includes. Mind you, the old code lacked `#pragma once`, so there may have been a reason it was commented out and so only included once.

Of course we can also switch to a forward declare for this particular include. Though we may want to do that separately, and verify includes are appropriate for switching to forward declares.